fix: update sensor wrongly reporting new update
An issue reported the update sensor wrongly indicating availability of a newer
version. The newer version in fact was older and marked as "testing". Since we
do not do a semantic version compare but just a simple string comparison, fix
by ignoring "testing" release. Also, make the whole logic a bit more robust.

resolves #7